requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
curlify==2.2.1
cycler==0.12.1
Flask==3.0.3
folium==0.17.0
fonttools==4.53.1
hyundai_kia_connect_api==3.22.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prometheus-client==0.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
xyzservices==2024.6.0
